Trait pallet_wnodl::pallet::Config

pub trait Config: Config {
    type Currency: ReservableCurrency<Self::AccountId>;
    type Event: From<Event<Self>> + IsType<<Self as Config>::Event>;
    type Oracles: Contains<Self::AccountId>;
    type KnownCustomers: Contains<Self::AccountId>;
    type ReserveTreasurerOrigin: EnsureOrigin<<Self as Config>::Origin>;
    type Reserve: OnUnbalanced<<<Self as Config>::Currency as Currency<<Self as Config>::AccountId>>::NegativeImbalance> + WithAccountId<Self::AccountId>;
    type WeightInfo: WeightInfo;
}

Configuration trait of this pallet.

Implement this type for a runtime in order to customize this pallet.

Associated Types

type Currency: ReservableCurrency<Self::AccountId>

type Event: From<Event<Self>> + IsType<<Self as Config>::Event>

type Oracles: Contains<Self::AccountId>

Trusted bots/oracles which can settle funds after wrapping is initiated

type KnownCustomers: Contains<Self::AccountId>

The customers who’ve gone under the KYC process and are eligible to wrap their NODLs

type ReserveTreasurerOrigin: EnsureOrigin<<Self as Config>::Origin>

Privileged origin that can initiate or settling wrapping funds from the reserve

type Reserve: OnUnbalanced<<<Self as Config>::Currency as Currency<<Self as Config>::AccountId>>::NegativeImbalance> + WithAccountId<Self::AccountId>

The chain’s reserve that is assigned to this pallet

type WeightInfo: WeightInfo

Weight information for extrinsics in this pallet.